backend/vendora/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
cloudinary==1.44.1
dataclasses-json==0.6.7
dj-database-url==3.0.0
Django==5.2.3
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-countries==7.6.1
django-extensions==4.1
django-filter==25.1
django-ordered-model==3.7.4
django-phonenumber-field==8.1.0
django-sslserver==0.22
djangorestframework==3.16.0
dnspython==2.7.0
drf-nested-routers==0.94.2
email_validator==2.2.0
extensions==0.4
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jsonpath-python==1.0.6
marshmallow==3.26.1
mypy_extensions==1.1.0
mysqlclient==2.2.7
narwhals==1.44.0
packaging==25.0
phonenumbers==9.0.7
pillow==11.2.1
plotly==6.2.0
psycopg2-binary==2.9.10
py3dbp==1.1.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2023.4
pyuploadcare==6.1.0
requests==2.32.4
shippo==3.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stripe==12.2.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
